// src/app/pages/menu-editor/config/entry-type.enum.ts
export enum EntryType {
    /** Gender */
    FIRST_LEVEL = 'fl',
    /** Category */
    SECOND_LEVEL = 'sl',
    /** Category, but link */
    SECOND_LEVEL_LINK = 'sll',
    /** Aka third level */
    LINK_BLOCK = 'lb',
    HIGHLIGHTED_LINK = 'hl',
    LINK = 'l',
    BANNER = 'b',
    WIDE_BANNER = 'wb',
    OPERATOR_LINK = 'o',
}